moved styles package to main
