+++ src/main/java/main/styles/KimStyle.java
@@ -1,23 +1,22 @@
-package styles;
+package main.styles;
 
 import java.awt.*;
 
 /**
- * The default style for the mancala board which uses brown colors to match
- * that of a wooden board.
+ * A style matching the colors of Kim's website for the mancala board
  * @author Jonathan Stewart Thomas
  * @version 1.0.0.230426
  */
-public class DefaultBoardStyle implements BoardStyle{
+public class KimStyle implements BoardStyle{
     private final Color boardColor, fontColor, pitColor;
 
     /**
-     * Sets the colors for the default style
+     * Sets the colors for the dark style
      */
-    public DefaultBoardStyle() {
-        boardColor = new Color(163, 114, 46);
-        fontColor = new Color(0,0,0);
-        pitColor = new Color(133, 91, 17);
+    public KimStyle() {
+        boardColor = new Color(50, 205, 50);
+        fontColor = new Color(0, 0, 238);
+        pitColor = new Color(255, 255, 102);
     }
 
     /**
